requirements.txt
AppKit==0.2.8
auto_mix_prep==0.2.0
Foundation==0.1.0a0.dev1
matplotlib==3.5.1
numpy==1.21.6
opencv_python==4.5.5.64
Pillow==9.1.0
PyQt4==4.11.4
PyQt5==5.15.6
PyScreeze==0.1.28
python_dateutil==2.8.2
qrcode==7.3.1
qtpy==2.0.1
Quartz==0.0.1.dev0
requests==2.27.1
setuptools==53.0.0
sphinx_rtd_theme==1.0.0
win32gui==221.6
WMI==1.4.9
Xlib==0.21
